renamed ProxyThread/Worker, made PTW subclass
ServerThread/ServerThreadWorker >>> ProxyThread/ProxyThreadWorker; PTW
now nested within PT
