+++ MultiTcp/src/main/java/edu/indiana/p538/AppConstants.java
@@ -12,4 +12,6 @@ public class AppConstants {
     public static final int MHEADER = 8;
     public static final int IP_LEN = 4;
     public static final int PORT_LEN = 2;
+    public static final int FIN_FLAG = 0x88;
+    public static final int RST_FLAG = 0x99;
 }
